flaskApp/requirements.txt
flask
requests
Flask-SQLAlchemy
flask-login
python-dotenv
Flask-Migrate
werkzeug
requests
openai
